+++ R/mol_parameters.R
@@ -92,8 +92,7 @@ get_m_cr_ch2_paras <- function(lw = NULL, lg = 0, ...) {
   if (is.null(lw)) lw = 2
   paras <- get_uncoupled_mol("-CrCH2", 3.913, "1H", -2, lw, lg,
                              "Inverted creatine CH2 group")
-  source <- "Proton NMR chemical shifts and coupling constants for brain
-              metabolites. NMR Biomed. 2000; 13:129-153."
+  source <- "Proton NMR chemical shifts and coupling constants for brain metabolites. NMR Biomed. 2000; 13:129-153."
   paras$source = source
   paras
 }
@@ -114,8 +113,7 @@ get_ala_paras <- function(lw = NULL, lg = 0, ...) {
                        j_coupling_mat = j_coupling_mat, scale_factor = 1,
                        lw = lw, lg = lg)
   
-  source <- "Proton NMR chemical shifts and coupling constants for brain
-              metabolites. NMR Biomed. 2000; 13: 129-153."
+  source <- "Proton NMR chemical shifts and coupling constants for brain metabolites. NMR Biomed. 2000; 13: 129-153."
   
   paras <- list(spin_groups = list(spin_group_a), name = "Ala", source = source,
                 full_name = "Alanine")
@@ -137,8 +135,7 @@ get_asc_paras <- function(lw = NULL, lg = 0, ...) {
                        j_coupling_mat = j_coupling_mat, scale_factor = 1,
                        lw = lw, lg = lg)
   
-  source <- "Detection of an antioxidant profile in the humain brain in vivo via
-              double editing with MEGA-PRESS. MRM. 2006; 56(6):1192-1199."
+  source <- "Detection of an antioxidant profile in the humain brain in vivo via double editing with MEGA-PRESS. MRM. 2006; 56(6):1192-1199."
   
   paras <- list(spin_groups = list(spin_group_a), name = "Asc", source = source,
                 full_name = "Ascorbate")
@@ -159,8 +156,7 @@ get_asp_paras <- function(lw = NULL, lg = 0, ...) {
                        j_coupling_mat = j_coupling_mat, scale_factor = 1,
                        lw = lw, lg = lg)
   
-  source <- "Proton NMR chemical shifts and coupling constants for brain
-              metabolites. NMR Biomed. 2000; 13:129-153."
+  source <- "Proton NMR chemical shifts and coupling constants for brain metabolites. NMR Biomed. 2000; 13:129-153."
   
   paras <- list(spin_groups = list(spin_group_a), name = "Asp", source = source,
                 full_name = "Aspartate")
@@ -194,8 +190,7 @@ get_cho_rt_paras <- function(lw = NULL, lg = 0, ...) {
                        j_coupling_mat = j_coupling_mat, scale_factor = 1,
                        lw = lw, lg = lg)
   
-  source <- "Corrigendum: Proton NMR chemical shifts and coupling constants for
-              brain metabolites. NMR Biomed. 2000; 13:129-153."
+  source <- "Corrigendum: Proton NMR chemical shifts and coupling constants for brain metabolites. NMR Biomed. 2000; 13:129-153."
   
   paras <- list(spin_groups = list(spin_group_a, spin_group_b), name = "Cho",
                 source = source, full_name = "Choline")
@@ -230,8 +225,7 @@ get_cho_paras <- function(lw = NULL, lg = 0, ...) {
                        j_coupling_mat = j_coupling_mat, scale_factor = 1,
                        lw = lw, lg = lg)
   
-  source <- "Corrigendum: Proton NMR chemical shifts and coupling constants for
-              brain metabolites. NMR Biomed. 2000; 13:129-153."
+  source <- "Corrigendum: Proton NMR chemical shifts and coupling constants for brain metabolites. NMR Biomed. 2000; 13:129-153."
   
   paras <- list(spin_groups = list(spin_group_a, spin_group_b), name = "Cho",
                 source = source, full_name = "Choline")
@@ -246,8 +240,7 @@ get_cr_paras <- function(lw = NULL, lg = 0, ...) {
   paras_b <- get_uncoupled_mol("Cr", 3.913, "1H", 2, lw, lg)
   paras$spin_groups[[2]] <- paras_b$spin_groups[[1]]
   
-  source <- "Proton NMR chemical shifts and coupling constants for brain
-              metabolites. NMR Biomed. 2000; 13:129-153."
+  source <- "Proton NMR chemical shifts and coupling constants for brain metabolites. NMR Biomed. 2000; 13:129-153."
   
   paras$source = source
   paras
@@ -258,8 +251,7 @@ get_cr_ch2_rt_paras <- function(lw = NULL, lg = 0, ...) {
   paras <- get_uncoupled_mol("CrCH2", 3.913, "1H", 2, lw, lg,
                              "Creatine CH2 group")
   
-  source <- "Proton NMR chemical shifts and coupling constants for brain
-              metabolites. NMR Biomed. 2000; 13:129-153."
+  source <- "Proton NMR chemical shifts and coupling constants for brain metabolites. NMR Biomed. 2000; 13:129-153."
   
   paras$source = source
   paras
@@ -270,8 +262,7 @@ get_cr_ch3_rt_paras <- function(lw = NULL, lg = 0, ...) {
   paras <- get_uncoupled_mol("CrCH3", 3.027, "1H", 3, lw, lg,
                              "Creatine CH2 group")
   
-  source <- "Proton NMR chemical shifts and coupling constants for brain
-              metabolites. NMR Biomed. 2000; 13:129-153."
+  source <- "Proton NMR chemical shifts and coupling constants for brain metabolites. NMR Biomed. 2000; 13:129-153."
   
   paras$source = source
   paras
@@ -298,8 +289,7 @@ get_gaba_paras <- function(lw = NULL, lg = 0, ...) {
                        j_coupling_mat = j_coupling_mat, scale_factor = 1,
                        lw = lw, lg = lg)
   
-  source <- "Corrigendum: Proton NMR chemical shifts and coupling constants for
-              brain metabolites. NMR Biomed. 2000; 13: 129-153."
+  source <- "Corrigendum: Proton NMR chemical shifts and coupling constants for brain metabolites. NMR Biomed. 2000; 13: 129-153."
   
   paras <- list(spin_groups = list(spin_group_a), name = "GABA",
                 source = source, full_name = "gamma-Aminobutyric acid")
@@ -328,9 +318,7 @@ get_gaba_rt_paras <- function(lw = NULL, lg = 0, ...) {
                        j_coupling_mat = j_coupling_mat, scale_factor = 1,
                        lw = lw, lg = lg)
   
-  source <- "Proton NMR chemical shifts and coupling constants for brain
-              metabolites. NMR Biomed. 2000; 13:129-153. Modified by MW for
-              room temperature phantom scans."
+  source <- "Proton NMR chemical shifts and coupling constants for brain metabolites. NMR Biomed. 2000; 13:129-153. Modified by MW for room temperature phantom scans."
   
   paras <- list(spin_groups = list(spin_group_a), name = "GABA",
                 source = source, full_name = "gamma-Aminobutyric acid")
@@ -359,8 +347,7 @@ get_gaba_jn_paras <- function(lw = NULL, lg = 0, ...) {
                        j_coupling_mat = j_coupling_mat, scale_factor = 1,
                        lw = lw, lg = lg)
   
-  source <- "J-Difference Editing of Gamma-Aminobutyric Acid (GABA): Simulated
-              and Experimental Multiplet Patterns. MRM 2013; 70:1183-1191."
+  source <- "J-Difference Editing of Gamma-Aminobutyric Acid (GABA): Simulated and Experimental Multiplet Patterns. MRM 2013; 70:1183-1191."
   
   paras <- list(spin_groups = list(spin_group_a), name = "GABA",
                 source = source, full_name = "gamma-Aminobutyric acid")
@@ -386,8 +373,7 @@ get_gln_paras <- function(lw = NULL, lg = 0, ...) {
                        j_coupling_mat = j_coupling_mat, scale_factor = 1,
                        lw = lw, lg = lg)
   
-  source <- "Proton NMR chemical shifts and coupling constants for brain
-              metabolites. NMR Biomed. 2000; 13:129-153."
+  source <- "Proton NMR chemical shifts and coupling constants for brain metabolites. NMR Biomed. 2000; 13:129-153."
   
   paras <- list(spin_groups = list(spin_group_a), name = "Gln",
                 source = source, full_name = "Glutamine")
@@ -429,8 +415,7 @@ get_gsh_paras <- function(lw = NULL, lg = 0, ...) {
                        j_coupling_mat = j_coupling_mat, scale_factor = 1,
                        lw = lw, lg = lg)
   
-  source <- "Proton NMR chemical shifts and coupling constants for brain
-              metabolites. NMR Biomed. 2000; 13:129-153."
+  source <- "Proton NMR chemical shifts and coupling constants for brain metabolites. NMR Biomed. 2000; 13:129-153."
   
   paras <- list(spin_groups = list(spin_group_a, spin_group_b, spin_group_c),
                 name = "GSH", source = source, full_name = "Glutathione")
@@ -443,8 +428,7 @@ get_gly_paras <- function(lw = NULL, lg = 0, ...) {
   if (is.null(lw)) lw = 2
   paras <- get_uncoupled_mol("Gly", 3.548, "1H", 2, lw, lg, "Glycine")
   
-  source <- "Proton NMR chemical shifts and coupling constants for brain
-              metabolites. NMR Biomed. 2000; 13:129-153."
+  source <- "Proton NMR chemical shifts and coupling constants for brain metabolites. NMR Biomed. 2000; 13:129-153."
   
   paras$source <- source
   paras
@@ -466,8 +450,7 @@ get_bhb_paras <- function(lw = NULL, lg = 0, ...) {
                        j_coupling_mat = j_coupling_mat, scale_factor = 1,
                        lw = lw, lg = lg)
   
-  source <- "In Vivo NMR Spectroscopy: Principles and Techniques,
-              Robin A. de Graaf"
+  source <- "In Vivo NMR Spectroscopy: Principles and Techniques, Robin A. de Graaf"
   
   paras <- list(spin_groups = list(spin_group_a), name = "BHB",
                 source = source, full_name = "beta-Hydroxybutyrate")
@@ -511,8 +494,7 @@ get_glyc_paras <- function(lw = NULL, lg = 0, ...) {
                        j_coupling_mat = j_coupling_mat, scale_factor = 1,
                        lw = lw, lg = lg)
   
-  source <- "In Vivo NMR Spectroscopy: Principles and Techniques,
-              Robin A. de Graaf"
+  source <- "In Vivo NMR Spectroscopy: Principles and Techniques, Robin A. de Graaf"
   
   paras <- list(spin_groups = list(spin_group_a), name = "Glyc",
                 source = source, full_name = "Glycerol")
@@ -536,8 +518,7 @@ get_ins_paras <- function(lw = NULL, lg = 0, ...) {
                        j_coupling_mat = j_coupling_mat, scale_factor = 1,
                        lw = lw, lg = lg)
   
-  source <- "Proton NMR chemical shifts and coupling constants for brain
-              metabolites. NMR Biomed. 2000; 13:129-153."
+  source <- "Proton NMR chemical shifts and coupling constants for brain metabolites. NMR Biomed. 2000; 13:129-153."
   
   paras <- list(spin_groups = list(spin_group_a), name = "Ins",
                 source = source, full_name = "myo-Inositol")
@@ -561,9 +542,7 @@ get_ins_rt_paras <- function(lw = NULL, lg = 0, ...) {
                        j_coupling_mat = j_coupling_mat, scale_factor = 1,
                        lw = lw, lg = lg)
   
-  source <- "Proton NMR chemical shifts and coupling constants for brain
-              metabolites. NMR Biomed. 2000; 13:129-153. Modified by MW for
-              room temperature phantom scans."
+  source <- "Proton NMR chemical shifts and coupling constants for brain metabolites. NMR Biomed. 2000; 13:129-153. Modified by MW for room temperature phantom scans."
   
   paras <- list(spin_groups = list(spin_group_a), name = "Ins", source = source,
                 full_name = "myo-Inositol")
@@ -706,9 +685,7 @@ get_naa_paras <- function(lw = NULL, lg = 0, ...) {
                        j_coupling_mat = j_coupling_mat, scale_factor = 1,
                        lw = lw, lg = lg)
   
-  source <- "Proton NMR chemical shifts and coupling constants for brain
-              metabolites. NMR Biomed. 2000; 13:129-153. Exlcuding the resonance
-              at 7.82 ppm."
+  source <- "Proton NMR chemical shifts and coupling constants for brain metabolites. NMR Biomed. 2000; 13:129-153. Exlcuding the resonance at 7.82 ppm."
   
   paras <- list(spin_groups = list(spin_group_a,spin_group_b), name = "NAA",
                 source = source, full_name = "N-acetylaspartate")
@@ -735,10 +712,7 @@ get_naa_rt_paras <- function(lw = NULL, lg = 0, ...) {
                        j_coupling_mat = j_coupling_mat, scale_factor = 1,
                        lw = lw, lg = lg)
   
-  source <- "Proton NMR chemical shifts and coupling constants for brain
-              metabolites. NMR Biomed. 2000; 13:129-153. Modified by MW for
-              room temperature phantom scans. Exlcuding the resonance at 7.82
-              ppm."
+  source <- "Proton NMR chemical shifts and coupling constants for brain metabolites. NMR Biomed. 2000; 13:129-153. Modified by MW for room temperature phantom scans. Exlcuding the resonance at 7.82 ppm."
   
   paras <- list(spin_groups = list(spin_group_a,spin_group_b), name = "NAA",
                 source = source, full_name = "N-acetylaspartate")
@@ -766,9 +740,7 @@ get_naa2_paras <- function(lw = NULL, lg = 0, ...) {
                        j_coupling_mat = j_coupling_mat, scale_factor = 1,
                        lw = lw, lg = lg)
   
-  source <- "Proton NMR chemical shifts and coupling constants for brain
-              metabolites. NMR Biomed. 2000; 13:129-153. Including the
-              resonance at 7.82 ppm."
+  source <- "Proton NMR chemical shifts and coupling constants for brain metabolites. NMR Biomed. 2000; 13:129-153. Including the resonance at 7.82 ppm."
   
   paras <- list(spin_groups = list(spin_group_a,spin_group_b), name = "NAA",
                 source = source, full_name = "N-acetylaspartate")
@@ -781,9 +753,7 @@ get_naag_ch3_paras <- function(lw = NULL, lg = 0, ...) {
   paras <- get_uncoupled_mol("NAAG", 2.042, "1H", 3, lw, lg,
                              "N-acetylaspartylglutamate")
   
-  source <- "Proton NMR chemical shifts and coupling constants for brain
-              metabolites. NMR Biomed. 2000; 13:129-153. Note, only the acetyl
-              moiety is simualted."
+  source <- "Proton NMR chemical shifts and coupling constants for brain metabolites. NMR Biomed. 2000; 13:129-153. Note, only the acetyl moiety is simualted."
   
   paras$source <- source
   paras
@@ -824,11 +794,7 @@ get_naag_paras <- function(lw = NULL, lg = 0, ...) {
                        j_coupling_mat = j_coupling_mat, scale_factor = 1,
                        lw = lw, lg = lg)
   
-  source <- "Chemical shifts are from the Govindaraju paper, J-couplings are
-              from : Characterisation of the 1H and 13C NMR spectra of N-
-              acetylaspartylglutamate and its detection in urine from patients
-              with Canavan disease, J Pharm Biomed Anal. 2003 Mar
-              10;31(3):455-63."
+  source <- "Chemical shifts are from the Govindaraju paper, J-couplings are from : Characterisation of the 1H and 13C NMR spectra of N- acetylaspartylglutamate and its detection in urine from patients with Canavan disease, J Pharm Biomed Anal. 2003 Mar 10;31(3):455-63."
   
   paras <- list(spin_groups = list(spin_group_a, spin_group_b, spin_group_c),
                 name = "NAAG", source = source,
@@ -864,8 +830,7 @@ get_pch_paras <- function(lw = NULL, lg = 0, ...) {
                        j_coupling_mat = j_coupling_mat, scale_factor = 1,
                        lw = lw, lg = lg)
   
-  source <- "Corrigendum: Proton NMR chemical shifts and coupling constants for
-              brain metabolites. NMR Biomed. 2000; 13:129-153."
+  source <- "Corrigendum: Proton NMR chemical shifts and coupling constants for brain metabolites. NMR Biomed. 2000; 13:129-153."
   
   paras <- list(spin_groups = list(spin_group_a, spin_group_b), name = "PCh",
                 source = source, full_name = "Phosphocholine")
@@ -881,8 +846,7 @@ get_pcr_paras <- function(lw = NULL, lg = 0, ...) {
   paras_b <- get_uncoupled_mol("PCr", 3.930, "1H", 2, lw, lg)
   paras$spin_groups[[2]] <- paras_b$spin_groups[[1]]
   
-  source <- "Proton NMR chemical shifts and coupling constants for brain
-              metabolites. NMR Biomed. 2000; 13:129-153."
+  source <- "Proton NMR chemical shifts and coupling constants for brain metabolites. NMR Biomed. 2000; 13:129-153."
   
   paras$source <- source
   paras
@@ -907,8 +871,7 @@ get_peth_paras <- function(lw = NULL, lg = 0, ...) {
                        j_coupling_mat = j_coupling_mat, scale_factor = 1,
                        lw = lw, lg = lg)
   
-  source <- "Corrigendum: Proton NMR chemical shifts and coupling constants for
-              brain metabolites. NMR Biomed. 2000; 13:129-153."
+  source <- "Corrigendum: Proton NMR chemical shifts and coupling constants for brain metabolites. NMR Biomed. 2000; 13:129-153."
   
   paras <- list(spin_groups = list(spin_group_a), name = "PEth",
                 source = source, full_name = "Phosphoethanolamine")
@@ -929,8 +892,7 @@ get_ser_paras <- function(lw = NULL, lg = 0, ...) {
                        j_coupling_mat = j_coupling_mat, scale_factor = 1,
                        lw = lw, lg = lg)
   
-  source <- "Corrigendum: Proton NMR chemical shifts and coupling constants for
-              brain metabolites. NMR Biomed. 2000; 13:129-153."
+  source <- "Corrigendum: Proton NMR chemical shifts and coupling constants for brain metabolites. NMR Biomed. 2000; 13:129-153."
   
   paras <- list(spin_groups = list(spin_group_a), name = "Ser", source = source,
                 full_name = "Serine")
@@ -942,8 +904,7 @@ get_sins_paras <- function(lw = NULL, lg = 0, ...) {
   if (is.null(lw)) lw = 2
   paras <- get_uncoupled_mol("sIns", 3.34, "1H", 6, lw, lg, "scyllo-Inositol")
   
-  source <- "Proton NMR chemical shifts and coupling constants for brain
-              metabolites. NMR Biomed. 2000; 13:129-153."
+  source <- "Proton NMR chemical shifts and coupling constants for brain metabolites. NMR Biomed. 2000; 13:129-153."
   
   paras$source <- source
   paras
@@ -964,10 +925,7 @@ get_msm_paras <- function(lw = NULL, lg = 0, ...) {
   paras <- get_uncoupled_mol("MSM", 3.142, "1H", 6, lw, lg,
                              "Methylsulfonylmethane")
   
-  source <- "Kaiser LG, Russell D, Maschmeyer T, Redfern RL, Inglis BA.
-    Methylsulfonylmethane (MSM): A chemical shift reference for 1 H MRS of human
-    brain. Magn Reson Med. 2020 Apr;83(4):1157-1167. doi: 10.1002/mrm.27997.
-    Epub 2019 Sep 30. PMID: 31566256."
+  source <- "Kaiser LG, Russell D, Maschmeyer T, Redfern RL, Inglis BA. Methylsulfonylmethane (MSM): A chemical shift reference for 1 H MRS of human brain. Magn Reson Med. 2020 Apr;83(4):1157-1167. doi: 10.1002/mrm.27997. Epub 2019 Sep 30. PMID: 31566256."
   
   paras$source <- source
   paras
@@ -978,8 +936,7 @@ get_ace_paras <- function(lw = NULL, lg = 0, ...) {
   paras <- get_uncoupled_mol("Ace", 2.222, "1H", 3, lw, lg,
                              "Acetone")
   
-  source <- "In Vivo NMR Spectroscopy Principles and Techniques, Robin A. de
-             Graaf, Third Edition, 2019, Wiley."
+  source <- "In Vivo NMR Spectroscopy Principles and Techniques, Robin A. de Graaf, Third Edition, 2019, Wiley."
   
   paras$source <- source
   paras
@@ -989,8 +946,7 @@ get_pyr_paras <- function(lw = NULL, lg = 0, ...) {
   if (is.null(lw)) lw = 2
   paras <- get_uncoupled_mol("Pyr", 2.3580, "1H", 3, lw, lg, "Pyruvate")
   
-  source <- "In Vivo NMR Spectroscopy Principles and Techniques, Robin A. de
-             Graaf, Third Edition, 2019, Wiley."
+  source <- "In Vivo NMR Spectroscopy Principles and Techniques, Robin A. de Graaf, Third Edition, 2019, Wiley."
   
   paras$source <- source
   paras
@@ -1000,8 +956,7 @@ get_suc_paras <- function(lw = NULL, lg = 0, ...) {
   if (is.null(lw)) lw = 2
   paras <- get_uncoupled_mol("Suc", 2.3920, "1H", 4, lw, lg, "Succinate")
   
-  source <- "Proton NMR chemical shifts and coupling constants for brain
-              metabolites. NMR Biomed. 2000; 13:129-153."
+  source <- "Proton NMR chemical shifts and coupling constants for brain metabolites. NMR Biomed. 2000; 13:129-153."
   
   paras$source <- source
   paras
@@ -1023,8 +978,7 @@ get_tau_paras <- function(lw = NULL, lg = 0, ...) {
                        j_coupling_mat = j_coupling_mat, scale_factor = 1,
                        lw = lw, lg = lg)
   
-  source <- "Corrigendum: Proton NMR chemical shifts and coupling constants for
-               brain metabolites. NMR Biomed. 2000; 13:129-153."
+  source <- "Corrigendum: Proton NMR chemical shifts and coupling constants for brain metabolites. NMR Biomed. 2000; 13:129-153."
   
   paras <- list(spin_groups = list(spin_group_a), name = "Tau", source = source,
                 full_name = "Taurine")
@@ -1044,8 +998,7 @@ get_thr_paras <- function(lw = NULL, lg = 0, ...) {
                        j_coupling_mat = j_coupling_mat, scale_factor = 1,
                        lw = lw, lg = lg)
   
-  source <- "In Vivo NMR Spectroscopy: Principles and Techniques,
-             Robin A. de Graaf"
+  source <- "In Vivo NMR Spectroscopy: Principles and Techniques, Robin A. de Graaf"
   
   paras <- list(spin_groups = list(spin_group_a), name = "Thr", source = source,
                 full_name = "Threonine")
@@ -1066,8 +1019,7 @@ get_lac_paras <- function(lw = NULL, lg = 0, ...) {
                        j_coupling_mat = j_coupling_mat, scale_factor = 1,
                        lw = lw, lg = lg)
   
-  source <- "Proton NMR chemical shifts and coupling constants for brain
-              metabolites. NMR Biomed. 2000; 13:129-153."
+  source <- "Proton NMR chemical shifts and coupling constants for brain metabolites. NMR Biomed. 2000; 13:129-153."
   
   paras <- list(spin_groups = list(spin_group_a), name = "Lac", source = source,
                 full_name = "Lactate")
@@ -1089,9 +1041,7 @@ get_lac_rt_paras <- function(lw = NULL, lg = 0, ...) {
                        j_coupling_mat = j_coupling_mat, scale_factor = 1,
                        lw = lw, lg = lg)
   
-  source <- "Proton NMR chemical shifts and coupling constants for brain
-             metabolites. NMR Biomed. 2000; 13:129-153. Modified by MW for
-             room temperature phantom scans."
+  source <- "Proton NMR chemical shifts and coupling constants for brain metabolites. NMR Biomed. 2000; 13:129-153. Modified by MW for room temperature phantom scans."
   
   paras <- list(spin_groups = list(spin_group_a), name = "Lac", source = source,
                 full_name = "Lactate")
@@ -1117,8 +1067,7 @@ get_glu_paras <- function(lw = NULL, lg = 0, ...) {
                        j_coupling_mat = j_coupling_mat, scale_factor = 1,
                        lw = lw, lg = lg)
   
-  source <- "Proton NMR chemical shifts and coupling constants for brain
-             metabolites. NMR Biomed. 2000; 13:129-153."
+  source <- "Proton NMR chemical shifts and coupling constants for brain metabolites. NMR Biomed. 2000; 13:129-153."
   
   paras <- list(spin_groups = list(spin_group_a), name = "Glu", source = source,
                 full_name = "Glutamate")
@@ -1144,9 +1093,7 @@ get_glu_rt_paras <- function(lw = NULL, lg = 0, ...) {
                        j_coupling_mat = j_coupling_mat, scale_factor = 1,
                        lw = lw, lg = lg)
   
-  source <- "Proton NMR chemical shifts and coupling constants for brain
-             metabolites. NMR Biomed. 2000; 13:129-153. Modified by MW for
-             room temperature phantom scans."
+  source <- "Proton NMR chemical shifts and coupling constants for brain metabolites. NMR Biomed. 2000; 13:129-153. Modified by MW for room temperature phantom scans."
   
   paras <- list(spin_groups = list(spin_group_a), name = "Glu", source = source,
                 full_name = "Glutamate")
@@ -1171,8 +1118,7 @@ get_a_glc_paras <- function(lw = NULL, lg = 0, ...) {
                        j_coupling_mat = j_coupling_mat, scale_factor = 1,
                        lw = lw, lg = lg)
   
-  source <- "Proton NMR chemical shifts and coupling constants for brain
-             metabolites. NMR Biomed. 2000; 13:129-153."
+  source <- "Proton NMR chemical shifts and coupling constants for brain metabolites. NMR Biomed. 2000; 13:129-153."
   
   paras <- list(spin_groups = list(spin_group_a), name = "aGlc",
                 source = source, full_name = "alpha-D-glucose")
@@ -1197,8 +1143,7 @@ get_b_glc_paras <- function(lw = NULL, lg = 0, ...) {
                        j_coupling_mat = j_coupling_mat, scale_factor = 1,
                        lw = lw, lg = lg)
   
-  source <- "Proton NMR chemical shifts and coupling constants for brain
-             metabolites. NMR Biomed. 2000; 13:129-153."
+  source <- "Proton NMR chemical shifts and coupling constants for brain metabolites. NMR Biomed. 2000; 13:129-153."
   
   paras <- list(spin_groups = list(spin_group_a), name = "bGlc",
                 source = source, full_name = "beta-D-glucose")
@@ -1238,9 +1183,7 @@ get_glc_paras <- function(lw = NULL, lg = 0, ...) {
                        j_coupling_mat = j_coupling_mat, scale_factor = 0.64,
                        lw = lw, lg = lg)
   
-  source <- "Proton NMR chemical shifts and coupling constants for brain
-             metabolites. NMR Biomed. 2000; 13:129-153. This is a combination
-             of alpha-glc (36%) and beta-glc (64%)."
+  source <- "Proton NMR chemical shifts and coupling constants for brain metabolites. NMR Biomed. 2000; 13:129-153. This is a combination of alpha-glc (36%) and beta-glc (64%)."
   
   paras <- list(spin_groups = list(spin_group_a, spin_group_b), name = "Glc",
                 source = source, full_name = "Glucose")
@@ -1289,8 +1232,7 @@ get_gpc_paras <- function(lw = NULL, lg = 0, ...) {
                        j_coupling_mat = j_coupling_mat, scale_factor = 1,
                        lw = lw, lg = lg)
   
-  source <- "Corrigendum: Proton NMR chemical shifts and coupling constants for
-             brain metabolites. NMR Biomed. 2000; 13:129-153."
+  source <- "Corrigendum: Proton NMR chemical shifts and coupling constants for brain metabolites. NMR Biomed. 2000; 13:129-153."
   
   paras <- list(spin_groups = list(spin_group_a, spin_group_b, spin_group_c),
                 name = "GPC", source = source,
@@ -1317,10 +1259,7 @@ get_2hg_paras <- function(lw = NULL, lg = 0, ...) {
                        j_coupling_mat = j_coupling_mat, scale_factor = 1,
                        lw = lw, lg = lg)
   
-  source <- "2-hydroxyglutarate detection by magnetic resonance spectroscopy
-             in subjects with IDH-mutated gliomas. Nat Med. 2012 Jan
-             26;18(4):624-9. Numbers are given in the supplementary
-             materials."
+  source <- "2-hydroxyglutarate detection by magnetic resonance spectroscopy in subjects with IDH-mutated gliomas. Nat Med. 2012 Jan 26;18(4):624-9. Numbers are given in the supplementary materials."
   
   paras <- list(spin_groups = list(spin_group_a), name = "2HG", source = source,
                 full_name = "2-hydroxyglutarate")
@@ -1339,12 +1278,7 @@ get_cit_paras <- function(lw = NULL, lg = 0, ...) {
                        j_coupling_mat = j_coupling_mat, scale_factor = 1,
                        lw = lw, lg = lg)
   
-  source <- "Optimizing PRESS localized citrate detection at 3 Tesla. Magn Reson
-             Med. 2005 Jul;54(1):51-8 and van der Graaf M, Heerschap A. Effect
-             of Cation Binding on the Proton Chemical Shifts and the Spin-Spin
-             Coupling Constant of Citrate. J Magn Reson B. 1996
-             Jul;112(1):58-62. Central frequency assumed to be 2.6 ppm, delta
-             chemical shift of 0.147 ppm and j-coupling value of 16.1 Hz."
+  source <- "Optimizing PRESS localized citrate detection at 3 Tesla. Magn Reson Med. 2005 Jul;54(1):51-8 and van der Graaf M, Heerschap A. Effect of Cation Binding on the Proton Chemical Shifts and the Spin-Spin Coupling Constant of Citrate. J Magn Reson B. 1996 Jul;112(1):58-62. Central frequency assumed to be 2.6 ppm, delta chemical shift of 0.147 ppm and j-coupling value of 16.1 Hz."
   
   paras <- list(spin_groups = list(spin_group_a), name = "Cit", source = source,
                 full_name = "Citrate")
@@ -1370,8 +1304,7 @@ get_val_paras <- function(lw = NULL, lg = 0, ...) {
                        j_coupling_mat = j_coupling_mat, scale_factor = 1,
                        lw = lw, lg = lg)
   
-  source <- "Proton NMR chemical shifts and coupling constants for brain
-             metabolites. NMR Biomed. 2000; 13:129-153."
+  source <- "Proton NMR chemical shifts and coupling constants for brain metabolites. NMR Biomed. 2000; 13:129-153."
   
   paras <- list(spin_groups = list(spin_group_a), name = "Val",
                 source = source, full_name = "Valine")